refactor: ExceptionHandlerMiddleware internal
WIP.

TODO: compatibility imports, deprecations.
