+++ litestar/middleware/_internal/exceptions/__init__.py
@@ -1,0 +1,3 @@
+from litestar.middleware._internal.exceptions.middleware import ExceptionHandlerMiddleware
+
+__all__ = ("ExceptionHandlerMiddleware",)
