+++ litestar/middleware/_internal/exceptions/templates/scripts.js
