+++ litestar/middleware/cors.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from litestar.middleware import _internal
+from litestar.middleware._internal import cors
 from litestar.utils.deprecation import warn_deprecation
 
 
@@ -15,5 +15,5 @@ def __getattr__(name: str) -> Any:
             removal_in="3.0",
             info="CORS middleware has been removed from the public API.",
         )
-        return _internal.CORSMiddleware
+        return cors.CORSMiddleware
     raise AttributeError(f"module {__name__} has no attribute {name}")
