--- litestar/middleware/exceptions/__init__.py
@@ -1,3 +1,0 @@
-from litestar.middleware.exceptions.middleware import ExceptionHandlerMiddleware
-
-__all__ = ("ExceptionHandlerMiddleware",)
