+++ litestar/security/session_auth/middleware.py
@@ -3,11 +3,11 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any, Awaitable, Callable, Sequence
 
 from litestar.exceptions import NotAuthorizedException
+from litestar.middleware._internal.exceptions import ExceptionHandlerMiddleware
 from litestar.middleware.authentication import (
     AbstractAuthenticationMiddleware,
     AuthenticationResult,
 )
-from litestar.middleware.exceptions import ExceptionHandlerMiddleware
 from litestar.types import Empty, Method, Scopes
 
 __all__ = ("MiddlewareWrapper", "SessionAuthMiddleware")
